@@ -1,6 +1,4 @@
-import Vue from 'vue'
-import App from './App.vue'
+import { createApp } from 'vue';
+import App from './App.vue';
 
-new Vue({
-  render: h => h(App)
-}).$mount('#app')
+createApp(App).mount('#app');
